move the dynamic logo into its own file
